Move twitter to separate module.
