+++ frontend/app/events/eventDetail/index.ts
@@ -1,2 +1,1 @@
 export { EventDetailComponent } from "./eventDetail.component"
-export { TweetComponent } from "./eventDetail.component"
