+++ frontend/app/twitter/index.ts
@@ -1,0 +1,1 @@
+export { TweetComponent } from "./tweet.component";
