+++ frontend/app/twitter/tweet.component.html
@@ -1,0 +1,1 @@
+<div [style.display]="displayStyle()" [innerHtml]="inner | async" #tweetWrapper></div>
