+++ gulp.config.js
@@ -97,16 +97,17 @@ module.exports = function () {
           '@angular/platform-browser-dynamic': { main: 'index.js', defaultExtension: 'js' },
           '@angular/router': { main: 'index.js', defaultExtension: 'js' },
           '@angular/upgrade': { main: 'index.js', defaultExtension: 'js' },
+          'app/about': { main: 'index.js'},
           'app/captcha': { main: 'index.js' },
           'app/contact': { main: 'index.js' },
           'app/events': { main: 'index.js' },
-          'app/events/shared': { main: 'index.js' },
-          'app/events/eventDetail': { main: 'index.js' },
           'app/events/event': { main: 'index.js' },
+          'app/events/eventDetail': { main: 'index.js' },
+          'app/events/shared': { main: 'index.js' },
           'app/facebook': { main: 'index.js' },
-          'app/map': { main: 'index.js'},
           'app/imprint': { main: 'index.js'},
-          'app/about': { main: 'index.js'},
+          'app/map': { main: 'index.js'},
+          'app/twitter': { main: 'index.js'},
       },
     };
