+++ systemjs.config.js
@@ -20,16 +20,17 @@
   };
 
   var appBarrels = [
+    'app/about',
     'app/captcha',
     'app/contact',
     'app/events',
-    'app/events/shared',
-    'app/events/eventDetail',
     'app/events/event',
+    'app/events/eventDetail',
+    'app/events/shared',
     'app/facebook',
-    'app/map',
     'app/imprint',
-    'app/about',
+    'app/map',
+    'app/twitter',
   ];
 
   appBarrels.forEach(function(barrelName) {
